requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.1
Pillow==12.0.0
python-decouple==3.8
requests==2.32.5
python-dotenv==1.2.1